patreon-webhook/infrastructure/requirements.txt
aws-cdk-lib>2.0,<3.0